Move imagetool tests to imagetool package
