+++ JHotDraw/src/test/java/imagetool/WhenImageAdded.java
@@ -1,9 +1,10 @@
-package org.jhotdraw.samples.svg.figures;
+package imagetool;
 
 import com.tngtech.jgiven.Stage;
 import com.tngtech.jgiven.annotation.ExpectedScenarioState;
 import org.jhotdraw.draw.DrawingEditor;
 import org.jhotdraw.samples.svg.SVGCreateFromFileTool;
+import org.jhotdraw.samples.svg.figures.SVGImageFigure;
 
 import static org.junit.Assert.assertNull;
