+++ JHotDraw/src/test/java/imagetool/test_image.jpg
